Revert "ioq3 sound"
This reverts commit 6e94bb6eca4adae410683dfbae13f4720a9e3aac.

Revert "Removed DebugGraph color param"

This reverts commit eef8f65830d8db3a8aaddee09d63509b004b9670.

Revert "New sound system"

This reverts commit ba0ed7504d290d58e912bcc49894653693a6135e.

Revert "ioq3 changes"

This reverts commit 41fc217a06021d0741662f7b6271d66f63f5fd11.

Revert "moves"

This reverts commit 02f22560792a810b30e88d95312bbfce0fa0bc03.

Revert "ioquake3 porting"

This reverts commit ca3340b1583fde34445fca970bf15043e2b634b7.

+++ code/sys/con_win32.c
@@ -139,7 +139,7 @@ static void CON_HistPrev( void )
 	qconsole_history_pos = pos;
 	Q_strncpyz( qconsole_line, qconsole_history[ qconsole_history_pos ], 
 		sizeof( qconsole_line ) );
-	qconsole_linelen = strlen( qconsole_line );
+	qconsole_linelen = (int)strlen( qconsole_line );
 	qconsole_cursor = qconsole_linelen;
 }
 
@@ -172,7 +172,7 @@ static void CON_HistNext( void )
 	qconsole_history_pos = pos;
 	Q_strncpyz( qconsole_line, qconsole_history[ qconsole_history_pos ],
 		sizeof( qconsole_line ) );
-	qconsole_linelen = strlen( qconsole_line );
+	qconsole_linelen = (int)strlen( qconsole_line );
 	qconsole_cursor = qconsole_linelen;
 }
 
@@ -416,7 +416,7 @@ char *CON_Input( void )
 			Field_AutoComplete( &f );
 			Q_strncpyz( qconsole_line, f.buffer,
 				sizeof( qconsole_line ) );
-			qconsole_linelen = strlen( qconsole_line );
+			qconsole_linelen = (int)strlen( qconsole_line );
 			qconsole_cursor = qconsole_linelen;
 			break;
 		}
